+++ app/Http/Controllers/IngredientController.php
@@ -17,7 +17,7 @@ use Kami\Cocktail\Repository\IngredientRepository;
 use Kami\Cocktail\Http\Resources\IngredientResource;
 use Kami\Cocktail\Http\Filters\IngredientQueryFilter;
 use Spatie\QueryBuilder\Exceptions\InvalidFilterQuery;
-use Kami\Cocktail\DataObjects\Ingredient\Ingredient as IngredientDTO;
+use Kami\Cocktail\DTO\Ingredient\Ingredient as IngredientDTO;
 
 class IngredientController extends Controller
 {