+++ app/Models/Cocktail.php
@@ -6,7 +6,6 @@ namespace Kami\Cocktail\Models;
 
 use Carbon\Carbon;
 use Kami\Cocktail\Utils;
-use Illuminate\Support\Str;
 use Laravel\Scout\Searchable;
 use Spatie\Sluggable\HasSlug;
 use Symfony\Component\Uid\Ulid;
@@ -235,91 +234,11 @@ class Cocktail extends Model implements UploadableInterface
             'name' => $this->name,
             'slug' => $this->slug,
             'description' => $this->description,
-            'image_url' => $this->getMainImageUrl(),
+            'image_url' => $this->getMainImageThumbUrl(),
             'short_ingredients' => $this->ingredients->pluck('ingredient.name'),
             'tags' => $this->tags->pluck('name'),
             'bar_id' => $this->bar_id,
         ];
-    }
-
-    public function share(bool $useUrls = false, bool $inlineImages = false): array
-    {
-        $data = [];
-        $cocktailId = Str::slug($this->name);
-
-        $data['_id'] = $cocktailId;
-        $data['name'] = $this->name;
-        $data['instructions'] = $this->instructions;
-        $data['description'] = $this->description ?? null;
-        $data['garnish'] = $this->garnish;
-        $data['source'] = $this->source;
-        $data['tags'] = $this->tags->pluck('name')->toArray();
-        $data['abv'] = $this->abv;
-        $data['created_at'] = $this->created_at;
-        $data['updated_at'] = $this->updated_at;
-
-        if ($this->glass_id) {
-            $data['glass'] = $this->glass->name;
-        }
-
-        if ($this->cocktail_method_id) {
-            $data['method'] = $this->method->name;
-        }
-
-        $data['ingredients'] = $this->ingredients->map(function (CocktailIngredient $cIngredient) {
-            $ingredient = [];
-            $ingredient['_id'] = Str::slug($cIngredient->ingredient->name);
-            $ingredient['sort'] = $cIngredient->sort ?? 0;
-            $ingredient['name'] = $cIngredient->ingredient->name;
-            $ingredient['amount'] = $cIngredient->amount;
-            if ($cIngredient->amount_max) {
-                $ingredient['amount_max'] = $cIngredient->amount_max;
-            }
-            $ingredient['units'] = $cIngredient->units;
-            if ($cIngredient->note) {
-                $ingredient['note'] = $cIngredient->note;
-            }
-            if ((bool) $cIngredient->optional === true) {
-                $ingredient['optional'] = (bool) $cIngredient->optional;
-            }
-
-            if ($cIngredient->substitutes->isNotEmpty()) {
-                $ingredient['substitutes'] = $cIngredient->substitutes->map(function (CocktailIngredientSubstitute $substitute) {
-                    return [
-                        '_id' => Str::slug($substitute->ingredient->name),
-                        'name' => $substitute->ingredient->name,
-                        'amount' => $substitute->amount,
-                        'amount_max' => $substitute->amount_max,
-                        'units' => $substitute->units,
-                    ];
-                })->toArray();
-            }
-
-            return $ingredient;
-        })->toArray();
-
-        if ($this->utensils->isNotEmpty()) {
-            $data['utensils'] = $this->utensils->pluck('name')->toArray();
-        }
-
-        $data['images'] = $this->images->map(function (Image $image, int $key) use ($cocktailId, $useUrls) {
-            $img = [
-                'sort' => $image->sort,
-                'placeholder_hash' => $image->placeholder_hash,
-                'copyright' => $image->copyright,
-            ];
-
-            // @deprecated move everything to url
-            if ($useUrls) {
-                $img['url'] = $image->getImageUrl();
-            } else {
-                $img['file_name'] = $cocktailId . '-' . ($key + 1) . '.' . $image->file_extension;
-            }
-
-            return $img;
-        })->toArray();
-
-        return $data;
     }
 
     public function toText(): string