+++ app/Models/Image.php
@@ -4,13 +4,12 @@ declare(strict_types=1);
 
 namespace Kami\Cocktail\Models;
 
+use Intervention\Image\ImageManager;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Database\Eloquent\Collection;
-use Intervention\Image\Image as InterventionImage;
 use Illuminate\Database\Eloquent\Relations\MorphTo;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Intervention\Image\Facades\Image as ImageProcessor;
 
 class Image extends Model
 {
@@ -38,19 +37,24 @@ class Image extends Model
         return $disk->url($this->file_path);
     }
 
+    public function getImageDataURI(): ?string
+    {
+        if (!$this->file_path) {
+            return null;
+        }
+
+        $manager = ImageManager::imagick();
+        $disk = Storage::disk('uploads');
+
+        return $manager->read($disk->path($this->file_path))->encode()->toDataUri();
+    }
+
     /**
      * @return MorphTo<Ingredient|Cocktail|Model, Image>
      */
     public function imageable(): MorphTo
     {
         return $this->morphTo();
-    }
-
-    public function asInterventionImage(): InterventionImage
-    {
-        $disk = Storage::disk('uploads');
-
-        return ImageProcessor::make($disk->path($this->file_path));
     }
 
     public function getPath(): string