+++ tests/Feature/Import/FromCollectionTest.php
@@ -7,9 +7,9 @@ namespace Tests\Feature\Import;
 use Tests\TestCase;
 use Kami\Cocktail\Models\User;
 use Kami\Cocktail\Models\Cocktail;
-use Kami\Cocktail\Import\FromCollection;
-use Kami\Cocktail\Import\DuplicateActionsEnum;
+use Kami\Cocktail\External\Import\FromCollection;
 use Illuminate\Foundation\Testing\RefreshDatabase;
+use Kami\Cocktail\External\Import\DuplicateActionsEnum;
 
 class FromCollectionTest extends TestCase
 {