+++ tests/Unit/Services/ImageServiceTest.php
@@ -6,8 +6,8 @@ namespace Tests\Unit\Services;
 
 use Tests\TestCase;
 use Illuminate\Log\LogManager;
+use Intervention\Image\ImageManager;
 use Kami\Cocktail\Services\ImageService;
-use Intervention\Image\ImageManagerStatic;
 
 class ImageServiceTest extends TestCase
 {
@@ -16,7 +16,8 @@ class ImageServiceTest extends TestCase
         $logMock = $this->mock(LogManager::class);
         $imageService = new ImageService($logMock);
 
-        $interventionImageInstance = ImageManagerStatic::make(base_path('tests/fixtures/cocktail.jpg'));
+        $manager = ImageManager::imagick();
+        $interventionImageInstance = $manager->read(base_path('tests/fixtures/cocktail.jpg'));
 
         $hashKey = $imageService->generateThumbHash($interventionImageInstance);